+++ R/tecan_parse.R
@@ -18,15 +18,28 @@
 #'
 #' @export
 tecan_parse <- function(xlsx_file, xlsx_sheet = 1) {
-  # check parameters for validity
+  # check function arguments for validity
+
   if (!(is.character(xlsx_file))) {
-    stop("File path must be a non empty character")
+    # stop("File path must be a non empty character")
+    cli::cli_abort(c(
+            "{.var xlsx_file} must be a character",
+      "x" = "You've supplied a {.cls {class(xlsx_file)}}."
+    ))
   }
-  if (!(file.exists(xlsx_file))) {
-    stop(paste("File does not exist:", xlsx_file))
+
+  xlsx_file <- xlsx_file |> fs::as_fs_path()
+
+  if (!(fs::file_exists(xlsx_file))) {
+    cli::cli_abort(c(
+      "x" = "File does not exist: {.file {xlsx_file}}"
+    ))
   }
   if (!(is.numeric(xlsx_sheet))) {
-    stop("Sheet number must be numeric")
+    cli::cli_abort(c(
+            "{.var xlsx_sheet} must be numeric",
+      "x" = "You've supplied a {.cls {class(xlsx_sheet)}}."
+    ))
   }
 
   # import data from excel spreadsheet